backend/requirements-fastapi.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
PyJWT==2.8.0
opencv-python==4.10.0.84
mediapipe==0.10.14
numpy==1.26.4
email-validator==2.1.0
torch==2.4.1
torchvision==0.19.1

